refactor: 프로젝트 폴더 구조 및 코드 정리
