+++ src/features/tag/default.ts
@@ -1,0 +1,1 @@
+export const defaultTags: string[] = ["집중됨", "뿌듯함", "힘듦", "피곤함"];
