+++ src/features/tag/initialize.ts
@@ -1,0 +1,6 @@
+import { SetterOrUpdater } from "recoil";
+import { defaultTags } from "./default";
+
+export const initializeTagState = (set: SetterOrUpdater<string[]>) => {
+  set(defaultTags);
+};
