+++ src/hooks/useClickOutside.ts
@@ -1,4 +1,3 @@
-// hooks/useClickOutside.ts
 import { useEffect } from "react";
 
 export function useClickOutside<T extends HTMLElement>(
