+++ src/pages/LayoutWithoutFooter/CategoryManagePage.tsx
@@ -1,13 +1,12 @@
 import { useState } from "react";
 import { useRecoilState } from "recoil";
 import { useNavigate } from "react-router-dom";
-import { categoryState } from "../../store/categoryAtom";
-import { CategoryType } from "../../types/category";
+import { categoryState, CategoryType } from "../../features";
 import { FiPlus } from "react-icons/fi";
 import PageHeader from "../../components/PageHeader";
-import AddCategoryModal from "../../components/AddCategoryModal";
-import ManageItemList from "../../features/myPage/ManageItemList";
-import ColorPickerPopover from "../../components/ColorPickerPopover";
+import AddCategoryModal from "../../components/modals/AddCategoryModal";
+import ManageItemList from "../../components/features/myPage/ManageItemList";
+import ColorPickerPopover from "../../components/modals/ColorPickerPopover";
 
 export default function CategoryManagePage() {
   const navigate = useNavigate();
